# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LunamHub Family Command Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /family:
    get:
      operationId: listFamilyMembers
      tags: [family]
      summary: List all family members
      responses:
        "200":
          description: List of family members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Create a family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /family/{id}:
    get:
      operationId: getFamilyMember
      tags: [family]
      summary: Get a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
        "404":
          description: Not found
    patch:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Delete a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Calendar Events ──────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Chores ───────────────────────────────────────────────────────────────────
  /chores:
    get:
      operationId: listChores
      tags: [chores]
      summary: List chores
      parameters:
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Chores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chore"
    post:
      operationId: createChore
      tags: [chores]
      summary: Create a chore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
  /chores/{id}:
    get:
      operationId: getChore
      tags: [chores]
      summary: Get a chore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
        "404":
          description: Not found
    patch:
      operationId: updateChore
      tags: [chores]
      summary: Update a chore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
    delete:
      operationId: deleteChore
      tags: [chores]
      summary: Delete a chore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chores/{id}/complete:
    post:
      operationId: completeChore
      tags: [chores]
      summary: Mark a chore as completed (child action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
  /chores/{id}/approve:
    post:
      operationId: approveChore
      tags: [chores]
      summary: Approve a completed chore (parent action, awards points)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
...
components:
  schemas:
...
    FamilyMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
        role:
          type: string
          enum: [parent, child]
        pointsBalance:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - emoji
        - color
        - role
        - pointsBalance
        - createdAt
    FamilyMemberInput:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
        role:
          type: string
          enum: [parent, child]
      required:
        - name
        - emoji
        - color
        - role
    FamilyMemberUpdate:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        color:
          type: string
        role:
          type: string
          enum: [parent, child]
        pointsBalance:
          type: integer

    # ── Events ────────────────────────────────────────────────────────────────
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        allDay:
          type: boolean
        category:
          type: string
          enum: [school, sport, appointment, birthday, family, other]
        assignedMembers:
          type: array
          items:
            type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - date
        - allDay
        - category
        - assignedMembers
        - createdAt
    EventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        allDay:
          type: boolean
        category:
          type: string
          enum: [school, sport, appointment, birthday, family, other]
        assignedMembers:
          type: array
          items:
            type: integer
      required:
        - title
        - date
        - category
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        allDay:
          type: boolean
        category:
          type: string
          enum: [school, sport, appointment, birthday, family, other]
        assignedMembers:
          type: array
          items:
            type: integer

    # ── Chores ────────────────────────────────────────────────────────────────
    Chore:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedMember:
          $ref: "#/components/schemas/FamilyMember"
        dueDate:
          type: ["string", "null"]
        repeatType:
          type: string
          enum: [once, daily, weekly]
        pointsValue:
          type: integer
        status:
          type: string
          enum: [pending, completed, approved, missed]
        createdAt:
          type: string
      required:
        - id
        - title
        - repeatType
        - pointsValue
        - status
        - createdAt
    ChoreInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: integer
        dueDate:
          type: string
        repeatType:
          type: string
          enum: [once, daily, weekly]
        pointsValue:
          type: integer
      required:
        - title
        - repeatType
        - pointsValue
    ChoreUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        repeatType:
          type: string
          enum: [once, daily, weekly]
        pointsValue:
          type: integer
        status:
          type: string
          enum: [pending, completed, approved, missed]
...