# lib/api-zod/src/generated/types/choreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LunamHub Family Command Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { ChoreInputRepeatType } from './choreInputRepeatType';

export interface ChoreInput {
  title: string;
  description?: string;
  assignedTo?: number;
  dueDate?: string;
  repeatType: ChoreInputRepeatType;
  pointsValue: number;
}